requirements.txt
pycord
python-dotenv>=1.0.0
yt-dlp>=2024.4.9
requests>=2.31.0
urllib3>=2.0.7
aiohttp>=3.9.3
async-timeout>=4.0.3
typing-extensions>=4.9.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
openai-whisper>=20231117
PyNaCl>=1.5.0
ffmpeg-python>=0.2.0
edge-tts>=6.1.9
torch>=2.0.0
torchaudio>=2.0.0 